feat: VM삭제, 매개변수로 받아 화면에 표시하도록 수정
