+++ CleanArea/CleanArea/iOS/Main/ListView.swift
@@ -8,11 +8,9 @@
 import SwiftUI
 
 struct ListView: View {
-    @ObservedObject var vm = ListVM()
-    var tabType: TabType
+    var policyItems: [PolicyItem]
+        var tabType: TabType
     @State private var searchText = ""
-    @Environment(\.presentationMode) var presentationMode
-    
     
     var body: some View {
         NavigationView {
@@ -47,7 +45,7 @@ struct ListView: View {
                 List {
                     ForEach(currentItems, id: \.sequenceNumber) { item in
                         ZStack(alignment: .leading) {
-                            PolicyListItemView(sequenceNumber: item.sequenceNumber,
+                            ListItemView(sequenceNumber: item.sequenceNumber,
                                                polyBizSjnm: item.polyBizSjnm,
                                                progress: item.progress,
                                                bizPrdCn: item.bizPrdCn,
@@ -70,147 +68,15 @@ struct ListView: View {
     }
     
     private var currentItems: [PolicyItem] {
-        switch tabType {
-        case .hot:
-            return vm.popularPolicyItems
-        case .like:
-            return vm.likePolicyItems
-        case .recommand:
-            return vm.recommandPolicyItems.filter { item in
-                searchText.isEmpty || item.polyBizSjnm.localizedCaseInsensitiveContains(searchText)
-            }
-        }
-    }
-}
-
-struct PolicyListItemView: View {
-    var sequenceNumber: Int?
-    var polyBizSjnm: String
-    var progress: String
-    var bizPrdCn: String
-    var remainDate: String
-    var polyCategory: String
-
-    @State private var isLike: Bool = false
-
-
-    var body: some View {
-        HStack {
-            VStack(alignment: .leading) {
-                if (sequenceNumber != nil) {
-                    HStack{
-                        Text("\(sequenceNumber!)")
-                            .font(.system(size: 20))
-                        Text(polyBizSjnm)
-                            .font(.system(size: 20))
-                    }
-                } else {
-                    Text(polyBizSjnm)
-                        .font(.system(size: 20))
-                }
-                HStack {
-                    ZStack {
-                        RoundedRectangle(cornerRadius: 10)
-                            .fill(progressColor)
-                            .frame(width: 70, height: 25)
-
-                        Text(progressText)
-                            .font(.system(size: 16))
-                            .bold()
-                            .foregroundStyle(.white)
-
-                    }
-                    .padding(.leading, 10)
-
-                    VStack {
-                        Text(bizPrdCn)
-                            .font(.system(size: 10))
-                            .foregroundColor(.gray)
-                        Text(remainDate)
-                            .font(.system(size: 12))
-                            .foregroundColor(.black)
-                    }
+            switch tabType {
+            case .hot, .like:
+                return policyItems
+            case .recommand:
+                return policyItems.filter { item in
+                    searchText.isEmpty || item.polyBizSjnm.localizedCaseInsensitiveContains(searchText)
                 }
             }
-            Spacer()
-            
-            VStack {
-                Button(action: {
-                    isLike.toggle()
-                }) {
-                    Image(systemName: isLike ? "star.fill" : "star")
-                        .foregroundColor(.buttonGreen)
-                        .font(.system(size: 20))
-                }
-                .padding(2)
-                
-                Text(polyCategory)
-                    .font(.system(size: 15))
-                    .foregroundColor(.gray)
-            }
         }
-        .padding()
-        .background(Color.backgroundGreen)
-        .cornerRadius(10)
-        
-    }
-    
-    
-    private var progressColor: Color {
-        switch progress {
-        case "allTimes":
-            return Color.allTimes
-        case "deadline":
-            return Color.deadline
-        default:
-            return Color.proceed
-        }
-    }
-    
-    private var progressText: String {
-        switch progress {
-        case "allTimes":
-            return "상시"
-        case "deadline":
-            return "마감임박"
-        default:
-            return "진행중"
-        }
-    }
-}
-
-struct SearchBar: View {
-    @Binding var text: String
-    
-    var body: some View {
-        HStack {
-            TextField("정책이름을 검색해 주세요", text: $text)
-                .padding(7)
-                .padding(.horizontal, 25)
-                .background(Color(.systemGray6))
-                .cornerRadius(8)
-                .overlay(
-                    HStack {
-                        Image(systemName: "magnifyingglass")
-                            .foregroundColor(.gray)
-                            .frame(minWidth: 0, maxWidth: .infinity, alignment: .leading)
-                            .padding(.leading, 8)
-                        
-                        if text != "" {
-                            Button(action: {
-                                self.text = ""
-                                hideKeyboard()
-                            }) {
-                                Image(systemName: "multiply.circle.fill")
-                                    .foregroundColor(.gray)
-                                    .padding(.trailing, 8)
-                            }
-                        }
-                    }
-                )
-                .frame(width: 342)
-        }
-    }
 }
 
 extension View {
@@ -221,6 +87,8 @@ extension View {
 
 
 #Preview {
-    ListView(tabType: .hot)
+    ListView(policyItems: recommandPolicyItems, tabType: .hot)
 }
+
+
