+++ CleanArea/CleanArea/iOS/Main/SearchBar.swift
@@ -1,0 +1,43 @@
+//
+//  SearchBar.swift
+//  CleanArea
+//
+//  Created by 최동호 on 12/11/23.
+//
+
+import Foundation
+import SwiftUI
+
+struct SearchBar: View {
+    @Binding var text: String
+    
+    var body: some View {
+        HStack {
+            TextField("정책이름을 검색해 주세요", text: $text)
+                .padding(7)
+                .padding(.horizontal, 25)
+                .background(Color(.systemGray6))
+                .cornerRadius(8)
+                .overlay(
+                    HStack {
+                        Image(systemName: "magnifyingglass")
+                            .foregroundColor(.gray)
+                            .frame(minWidth: 0, maxWidth: .infinity, alignment: .leading)
+                            .padding(.leading, 8)
+                        
+                        if text != "" {
+                            Button(action: {
+                                self.text = ""
+                                hideKeyboard()
+                            }) {
+                                Image(systemName: "multiply.circle.fill")
+                                    .foregroundColor(.gray)
+                                    .padding(.trailing, 8)
+                            }
+                        }
+                    }
+                )
+                .frame(width: 342)
+        }
+    }
+}
